+++ src/language/sql/translate.rs
@@ -166,7 +166,7 @@ fn handle_table(identifier: SqlIdentifier) -> Result<MaybeAliasAlg, String> {
     let scan = match names.remove(0) {
         name if name.starts_with('$') => name.strip_prefix('$')
             .ok_or("Prefix not found".to_string())
-            .and_then(|rest| rest.parse::<i64>().map_err(|_| "Could not parse number".to_string()))
+            .and_then(|rest| rest.parse::<usize>().map_err(|_| "Could not parse number".to_string()))
             .map(|num| IndexScan(algebra::IndexScan::new(num)))?,
         name => TableScan(algebra::TableScan::new(name)),
     };
